refactor[#107] : SelectToggle , widgets 세그먼트로 이동
