+++ blog/src/slots/write/ui/ArticleWrite.tsx
@@ -6,7 +6,8 @@ import {
   createArticleWriteStore,
   useArticleWriteStore
 } from "../model";
-import { SeriesSelectToggle, TagSelectToggle } from "./items";
+import { SeriesSelectToggle } from "./SeriesSelectToggle";
+import { TagSelectToggle } from "./TagSelectToggle";
 import { useQueryClient } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
 import React, { useContext, useEffect, useRef } from "react";
