+++ blog/src/slots/write/ui/SeriesSelectToggle.tsx
