--- blog/src/slots/write/ui/items/index.ts
@@ -1,2 +1,0 @@
-export * from "./SeriesSelectToggle";
-export * from "./TagSelectToggle";
